WIP: Upgrade to the latest versions of Primus and UWS
It’s not clear that the previous version of UWS supported Node 8
https://github.com/uNetworking/uWebSockets/issues/630#issuecomment-309284379
and then even after upgrading that, as soon as the client sent a message to
the server, the websocket would disconnect, until Primus was upgraded.

WIP because publication-client’s Primus implementation might need to be
regenerated (though it seems to work).