Upgrade Primus and uws for node 10 support
Previously, this library used uws@0.14.3 which didn't have bindings for
linux 64 or darwin 64 ABIs. Prior to node 10, this wasn't a problem
since `process.versions.modules` did not support 64 ABIs. However,
with upgrades to node 10 and V8, 64 ABIs are now supported.

The latest releases of `uws` (10.148.1) has 64-bit bindings, and the
latest version of `primus` (7.3.3) supports the latest version of uws
(based on their dev depencencies).

Looking at the latest version of primus, it looks like they changed 2
options:
* 1 when initializing the server which is inaccessible outside this
library
* 1 when initializing the client which _is_ accessible outside the
library

The latter only applies to `publication-client` so we'll have to make a
major version change there, but `publication-server` can be
transparently upgraded.